Extracted useFindDuplicateRecordsQuery
